requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
numpy<2
pandas
faiss-cpu>=1.9.0
rank_bm25
sentence-transformers
torch
transformers
openai
python-dotenv
pyyaml
requests
beautifulsoup4
pdfplumber
python-docx
tiktoken
prometheus_client
pytest
httpx
# approximate for ColBERT if needed, widely used wrapper
# ragatouille 
# For now, we will stick to sentence-transformers for cross-encoder which is lighter and very effective.
# If full ColBERT is strictly needed we can add it later.
Faker
datasets<3.0.0
pyarrow
pinecone>=3.0.0
huggingface_hub
